requirements.txt
docker==4.2.2
loguru==0.5.1
Jinja2==2.11.2
PyYAML==5.3.1